split CustomMessages and CustomLogo into separate components from appearance settings
